+++ pangea-sdk/service/ai_guard/service.go
@@ -10,6 +10,8 @@ import (
 type Client interface {
 	GuardText(ctx context.Context, input *TextGuardRequest) (*pangea.PangeaResponse[TextGuardResult], error)
 	GuardTextWithRelevantContent(ctx context.Context, input *TextGuardRequest) (*pangea.PangeaResponse[TextGuardResult], error)
+	Guard(ctx context.Context, input GuardRequest) (*pangea.PangeaResponse[GuardResult], error)
+	GuardAsync(ctx context.Context, input GuardRequest) (*pangea.PangeaResponse[GuardResult], error)
 
 	// Base service methods.
 	pangea.BaseServicer